requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
httpx
pytest
pytest-asyncio
pytest-cov
requests
agents
openai
